refactor library
